codebuddy/pyproject.toml
[tool.poetry]
name = "codebuddy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.6"
pygithub = "^1.58.2"
flask = "^2.3.2"
tiktoken = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
